requirements.txt
flask==2.3.3
pillow==10.0.1
numpy==1.24.3
requests==2.31.0
gunicorn==21.2.0
